+++ libclang-bindings/test/test-clang-bindings.hs
@@ -1,5 +1,6 @@
 module Main (main) where
 
+import Test.Discover qualified as Discover
 import Test.Meta.IsConcrete qualified as IsConcrete
 import Test.Tasty
 import Test.Test.Exceptions qualified as Exceptions
@@ -12,6 +13,7 @@ import Test.Version qualified as Version
 main :: IO ()
 main = defaultMain $ testGroup "test-clang-bindings" [
       Version.tests
+    , Discover.tests
     , testGroup "Meta" [ -- Tests of the test infrastructure
           IsConcrete.tests
         ]